pyproject.toml
[tool.poetry]
name = "lumber-chunker"
version = "0.1.0"
description = ""
readme = "README.md"
packages = [{ include = "lumber_chunker" }]

[tool.poetry.dependencies]
python = "^3.11"
langchain-text-splitters = "0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
